Share the read-only SQLite reader across session adapters
AntiGravity's conversation databases and Cursor's agent stores pose the
same problem: another process holds the file open with a live WAL, so
`immutable=1` would read a torn snapshot and a plain open can fail. The
open / snapshot / retry strategy that solved it for AntiGravity is now
`LiveSQLiteReader`, leaving `AntigravityLiveSQLite` as just the two
queries that know that schema.

No behavior change — the reader is byte-for-byte the code it replaces.
